+++ examples/resources/nxos_bgp_instance/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_bgp_instance.example "sys/bgp/inst"