+++ examples/resources/nxos_bgp_instance/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_bgp_instance" "example" {
+  admin_state = "enabled"
+  asn         = "65001"
+}